requirements.txt
python-dotenv
pydantic
whisper
torch